+++ WalletKitCore/src/crypto/handlers/xtz/BRCryptoWalletManagerXTZ.c
@@ -295,7 +295,7 @@ cryptoWalletManagerRecoverFeeBasisFromFeeEstimateXTZ (BRCryptoWalletManager cwm,
     storageUsed = (int64_t)(storageUsed * 1.1);
     BRTezosUnitMutez mutezPerKByte = tezosMutezCreate (networkFee->pricePerCostFactor); // given as nanotez/byte
     // get the serialized txn size from the estimation payload
-    size_t sizeInKBytes = cryptoFeeBasisCoerceXTZ(initialFeeBasis)->xtzFeeBasis.u.estimate.sizeInKBytes;
+    double sizeInKBytes = cryptoFeeBasisCoerceXTZ(initialFeeBasis)->xtzFeeBasis.u.estimate.sizeInKBytes;
     
     BRTezosFeeBasis feeBasis = tezosFeeBasisCreateEstimate (mutezPerKByte,
                                                             sizeInKBytes,
